requirements.txt
tensorflow >= 2.6.0
numpy >= 1.22.3
flask >= 2.1.2
gunicorn >= 20.1.0
pillow